Fix issue #206 and issue #246: VTP problems
Issue 206: VECTO did not allow VTP with old (v1 & v2) vehicle xmls.
Also, VTP failed (in Declaration mode) when in VDRI torque converter
was active in gears > 1.

Solution 206: Old vehicle xmls are allowed only for VTP jobs. The
torque converter problem was fixed by copying the code from the
engineering mode which works properly.

Issue 246: Required info for the VTP report such as: the CO2 values,
Gearbox info and Tyre digest values in the MRF cannot be read because
the XSD has changed but the code that parses the MRF has not been
updated accordingly.

Solution 246: The code that reads the MRF has been updated to be able
to read properly both older and current (v0.9) MRFs.