fix #427: correct calculation of vehicle max speed
For the calculation of MaxVehicleSpeed, the N95 speed of the engine is
used. In some edge cases though, the speed of the vehicle may result to
exceeding the max speed of the last gear (i.e. manifested at coach
cycle on overspeed at about 102km/h). Therefore we need to use the
minimum of N95 and maxGBspeed @ last gear.