1Cto1L condition change and increase on coefficient applied to original DeltaDragLoad of TorqueConverter
- fix issue #542

- Error observed: brake torque determination method converges to a value
fitting tolerances, but the effective request tolerance is exceeded and
result in very small underload that leads to a crash
the change of criterion in brake torque determination method

- This is due to the fact that DryRun request of brake torque determination
method send back a DeltaDrag located at the torque converter output
level when overload or underload (instead of the ICE output when an ICE
operating point is reached). This deltaDrag is used as criteria to determine
the brake torque. To avoid false positive (point valid on Torque converter
level, but not on ICE level), and keep tendencies in the model this
DatadeltaDrag is multiplied by 10.

- Increasing this factor by 10% (from 10 to 11) is sufficient to avoid
the error, as well as increasing the criterion in Driver/SearchBrakingPower:
This partially fix issue #542: the computation does not crash
for HeavyUrban P31SD LowLoading,but still does for Urban P31DD RefLoad
an unexpected behavior is observed: going downhill, strategy
does not engage from 1C to 1L when it could due to the fact
that acceleration is zero when gearshift timing is reached,
and prefers to go to neutral and brake, leading to ice idling
and overconsumption compared to other cases allowing upshift
because gearshift timing is reached during previous acceleration

- tightening AT disengage condition when Tice>0 and Tout<0 by imposing
that acceleration should be negative, strictly (instead of <=0) completely
fix issue #542: computation does not crash for HeavyUrban
P31SD LowLoading and for Urban P31DD RefLoad, while behavior in gear
selection works as expected

- the fix has been tested on different case of vehicles and previous
issues: no significant change observed in consumption results
(max -0.08%) except for the input file of issue 542 where consumption is
reduced for the cycles that previously crashed (-0.91% for the HeavyUrban
and -2.42% for the Urban case)